Extract Java implementations to java package. Update docs
